% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicken-embryo.r
\docType{data}
\name{chicken.embryo}
\alias{chicken.embryo}
\title{Chicken Embryo Data}
\format{The format is:
List of 2: 
day: int [1:21] 1 2 3 4 5 6 7 8 9 10 ...; 
nT : int [1:21] 6 5 11 2 2 3 0 0 0 0 ...}
\source{
Jassim, E. W., Grossman, M., Koops, W. J. And Luykx, R. A. J. (1996). Multi-phasic analysis of embryonic
mortality in chickens. \emph{Poultry Sci.} 75, 464-71.
}
\usage{
data(chicken.embryo)
}
\description{
The chicken embryo dataset which contains \code{day}, number of days,  and \code{nT}, the corresponding frequencies.
}
\examples{
data(chicken.embryo)
}
\references{
Kuurman, W. W., Bailey, B. A., Koops, W. J. And Grossman, M. (2003). A model for failure of a chicken embryo to survive incubation. \emph{Poultry Sci.} 82, 214-22.

Guan, Z. (2017) Bernstein polynomial model for grouped continuous data.  \emph{Journal of Nonparametric Statistics}, 29(4):831-848.
}
\keyword{datasets}
