% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made.r
\name{chpt.exp}
\alias{chpt.exp}
\title{Exponential change-point}
\usage{
chpt.exp(x)
}
\arguments{
\item{x}{a vector of nondecreasing values of log-likelihood or -log of distance}
}
\value{
a list of exponential change-point, its p-value, and the likelihood ratios 
          of the exponential change-point model
}
\description{
Exponential change-point
}
