% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable-copula.r
\name{corr.hellinger}
\alias{corr.hellinger}
\title{Bhattacharyya coefficient and Hellinger correlation}
\usage{
corr.hellinger(dcopula, ...)
}
\arguments{
\item{dcopula}{a function object defining a 2d copula density function}

\item{...}{argument(s) of copula density function}
}
\value{
Bhattacharyya coefficient \code{B} and Hellinger correlation \code{eta}
}
\description{
Bhattacharyya coefficient and Hellinger correlation
}
\references{
Geenens, G. and Lafaye de Micheaux, P. (2022). The Hellinger correlation. 
    Journal of the American Statistical Association 117(538), 639–653.
}
