% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmesis-data.r
\docType{data}
\name{cosmesis}
\alias{cosmesis}
\title{Breast cosmesis data}
\format{
A data frame with 94 observations on the following 3 variables. 

\itemize{
  \item left left endpoint of the censoring interval in months
  \item right right endpoint of the censoring interval in months
  \item treat a factor with levels \code{RT} and \code{RCT} representing radiotherapy-only 
         and radiation plus chemotherapy treatments, respectively
}
}
\source{
Finkelstein, D. M. and Wolfe, R. A. (1985) 
    A semiparametric model for regression analysis of interval-censored failure time data. 
    Biometrics 41, 933–945.
}
\usage{
data(cosmesis)
}
\description{
Data contain the interval-censored times to cosmetic deterioration  
    for breast cancer patients undergoing radiation or radiation plus chemotherapy.
}
\examples{
data(cosmesis)
}
\references{
Finkelstein, D. M. (1986) A proportional hazards model for interval-censored 
   failure time data. Biometrics 42, 845–854.
}
\keyword{datasets}
