% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable-dr.r
\name{maple.dr}
\alias{maple.dr}
\title{Maximum approximate profile likelihood estimate of the density ratio model}
\usage{
maple.dr(
  x,
  y,
  M,
  regr,
  ...,
  interval = c(0, 1),
  alpha = NULL,
  vb = 0,
  controls = mable.ctrl(),
  progress = TRUE,
  message = TRUE
)
}
\arguments{
\item{x, y}{original two sample raw data, code{x}:"Control", \code{y}: "Case".}

\item{M}{a positive integer or a vector \code{(m0, m1)}.}

\item{regr}{regressor vector function \eqn{r(x)=(1,r_1(x),...,r_d(x))} 
which returns n x (d+1) matrix, n=length(x)}

\item{...}{additional arguments to be passed to regr}

\item{interval}{a vector \code{(a,b)} containing the endpoints of 
supporting/truncation interval of x and y.}

\item{alpha}{a given regression coefficient, missing value is imputed by 
logistic regression}

\item{vb}{code for vanishing boundary constraints, -1: f0(a)=0 only, 
1: f0(b)=0 only, 2: both, 0: none (default).}

\item{controls}{Object of class \code{mable.ctrl()} specifying iteration limit
and the convergence criterion for EM and Newton iterations. Default is 
\code{\link{mable.ctrl}}. See Details.}

\item{progress}{logical: should a text progressbar be displayed}

\item{message}{logical: should warning messages be displayed}
}
\value{
A list with components
\itemize{
  \item \code{m} the given or a selected degree by method of change-point
  \item \code{p} the estimated vector of mixture proportions \eqn{p = (p_0, \ldots, p_m)}
      with the given or selected degree \code{m}
  \item \code{alpha} the given regression coefficients
  \item \code{mloglik}  the maximum log-likelihood at degree \code{m}
  \item \code{interval} support/truncation interval \code{(a,b)}
  \item \code{M} the vector \code{(m0, m1)}, where \code{m1}, if greater than \code{m0}, is the
     largest candidate when the search stoped
  \item \code{lk} log-likelihoods evaluated at \eqn{m \in \{m_0, \ldots, m_1\}}
  \item \code{lr} likelihood ratios for change-points evaluated at \eqn{m \in \{m_0+1, \ldots, m_1\}}
  \item \code{pval} the p-values of the change-point tests for choosing optimal model degree
  \item \code{chpts} the change-points chosen with the given candidate model degrees
}
}
\description{
Select optimal degree with a given regression coefficients.
}
\details{
Suppose that ("control") and \code{y} ("case") are independent samples from   
f0 and f1 which satisfy f1(x)=f0(x)exp[alpha0+alpha'r(x)] 
with r(x)=(r1(x),...,r_d(x)). Maximum 
approximate Bernstein/Beta likelihood estimates of  f0 and f1 are calculated 
with a given regression coefficients which are efficient estimates  provided 
by other semiparametric methods such as logistic regression.
If support is (a,b) then replace r(x) by r[a+(b-a)x].
For a fixed \code{m}, using the Bernstein polynomial model for baseline \eqn{f_0},
MABLEs of \eqn{f_0} and parameters alpha can be estimated by EM algorithm and Newton  
iteration. If estimated lower bound \eqn{m_b} for \code{m} based on \code{y}
is smaller that that based on \code{x}, then switch \code{x} and \code{y} and
\eqn{f_1} is used as baseline. If \code{M=m} or \code{m0=m1=m}, then \code{m} is a 
preselected degree. If \code{m0<m1} it specifies the set of consective  
candidate model degrees \code{m0:m1} for searching an optimal degree by 
the change-point method, where \code{m1-m0>3}.
}
\references{
Guan, Z., Application of Bernstein Polynomial Model to Density 
 and ROC Estimation in a Semiparametric Density Ratio Model
}
\author{
Zhong Guan <zguan@iusb.edu>
}
