% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable-multivar.r
\name{marginal.p}
\alias{marginal.p}
\title{The mixing proportions of marginal distribution from the mixture of 
multivariate beta distribution}
\usage{
marginal.p(p, m)
}
\arguments{
\item{p}{the mixing proportions of the mixture of multivariate beta distribution}

\item{m}{the model  degrees \code{m=(m1,...,md)} of the mixture of 
multivariate beta distribution}
}
\value{
a list of mixing proportions of all the marginal distributions
}
\description{
The mixing proportions of marginal distribution from the mixture of 
multivariate beta distribution
}
