% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made.r
\name{mvecdf}
\alias{mvecdf}
\title{Multivariate empirical cumulative distribution evaluated at sample data}
\usage{
mvecdf(x)
}
\arguments{
\item{x}{an \code{n x d} matrix of data values, 
rows are \code{n} observations of \eqn{X=(X_1,...,X_d)}}
}
\value{
a vector of \code{n} values
}
\description{
Multivariate empirical cumulative distribution evaluated at sample data
}
