% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made.r
\name{mvpbeta}
\alias{mvpbeta}
\title{Component Beta cumulative distribution functions of the Bernstein polynomial model}
\usage{
mvpbeta(x, m)
}
\arguments{
\item{x}{\code{n x d} matrix, rows are n observations of X=(X1,...,Xd)}

\item{m}{vector of d nonnegative integers \code{m=(m[1],...,m[d])}.}
}
\value{
an \code{n x K} matrix, \code{K=(m[1]+1)...(m[d]+1)}.
}
\description{
Component Beta cumulative distribution functions of the Bernstein polynomial model
}
