% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.r
\name{plot.mable}
\alias{plot.mable}
\title{Plot mathod for class 'mable'}
\usage{
\method{plot}{mable}(
  x,
  which = c("density", "cumulative", "survival", "likelihood", "change-point", "all"),
  add = FALSE,
  lgd.x = NULL,
  lgd.y = NULL,
  nx = 512,
  ...
)
}
\arguments{
\item{x}{Class "mable" object return by \code{mablem}, \code{mable}, \code{mablem.group} or \code{mable.group} functions
which contains \code{p}, \code{mloglik}, and \code{M = m0:m1}, \code{lk}, \code{lr},}

\item{which}{indicates which graphs to plot, options are
"density", "cumulative", "likelihood", "change-point", "all". If not "all",
\code{which} can contain more than one options.}

\item{add}{logical add to an existing plot or not}

\item{lgd.x, lgd.y}{coordinates of position where the legend is displayed}

\item{nx}{number of evaluations of density, or cumulative distribution curve to be plotted.}

\item{...}{additional arguments to be passed to the base plot function}
}
\value{
The data used for 'plot()', 'lines()', or 'persp()' are returned invisibly.
}
\description{
Plot mathod for class 'mable'
}
