% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.r
\name{rmixbeta}
\alias{rmixbeta}
\title{Generate data from Bernstein polynomial model
Generate random sample from mixture of beta distributions with
 shapes \eqn{(i+1, m-i+1)}, \eqn{i=0,\ldots,m}, with given mixture
 proportions \eqn{p=(p_0, \ldots,p_m)}.}
\usage{
rmixbeta(n, p)
}
\arguments{
\item{n}{sample size}

\item{p}{a vector of mixture proportions, must contain at least four components}
}
\description{
Generate data from Bernstein polynomial model
Generate random sample from mixture of beta distributions with
 shapes \eqn{(i+1, m-i+1)}, \eqn{i=0,\ldots,m}, with given mixture
 proportions \eqn{p=(p_0, \ldots,p_m)}.
}
