% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.r
\name{summary.mable}
\alias{summary.mable}
\alias{summary.mable_reg}
\title{Summary mathods for classes 'mable' and 'mable_reg'}
\usage{
\method{summary}{mable}(object, ...)

\method{summary}{mable_reg}(object, ...)
}
\arguments{
\item{object}{Class "mable" or 'mable_reg' object return by \code{mable} or 
\code{mable.xxxx}  functions}

\item{...}{for future methods}
}
\value{
Invisibly returns its argument, \code{object}.
}
\description{
Produces a summary of a mable fit.
}
\examples{
\donttest{
## Breast Cosmesis Data
  bcos=cosmesis
  bcos2<-data.frame(bcos[,1:2], x=1*(bcos$treat=="RCT"))
  aft.res<-mable.aft(cbind(left, right)~x, data=bcos2, M=c(1, 30), g=.41, 
       tau=100, x0=1)
  summary(aft.res)
}
}
