% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made.r
\name{umc.mat}
\alias{umc.mat}
\title{Matrix of the uniform marginal constraints}
\usage{
umc.mat(m)
}
\arguments{
\item{m}{vector of d nonnegative integers \code{m=(m[1],...,m[d])}.}
}
\value{
an \code{|m| x K} matrix, \code{|m|=m[1]+...+m[d])}, \code{K=(m[1]+1)...(m[d]+1)}.
}
\description{
Matrix of the uniform marginal constraints
}
\details{
the matrix of the uniform marginal constraints \eqn{A} is used to form the 
    linear equality constraints on parameter \eqn{p}: \eqn{Ap=1/(m+1)}.
}
