\name{Vaal.Flow}
\alias{Vaal.Flow}
\docType{data}
\title{Vaal River Annual Flow Data}
\description{The annual flow data of Vaal River at Standerton as given by Table 1.1 of Linhart and Zucchini
(1986) give the flow in millions of cubic metres.}
\usage{data(Vaal.Flow)}
\format{
  The format is:
 int [1:65] 222 1094 452 1298 882 988 276 216 103 490 ...
}
\references{Linhart, H., and Zucchini, W. (1986), \emph{Model selection}, Wiley Series in Probability and Mathematical
Statistics: Applied Probability and Statistics, New York: John Wiley \& Sons Inc.
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Vaal.Flow)
## maybe str(Vaal.Flow) ; plot(Vaal.Flow) ...
}
\keyword{datasets}
