% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable-ph-model.r
\name{plot.mable_reg}
\alias{plot.mable_reg}
\title{Plot mathod for class 'mable_reg'}
\usage{
\method{plot}{mable_reg}(x, y, newdata = NULL, ntime = 512,
  xlab = "Time", which = c("survival", "likelihood", "change-point",
  "density", "all"), add = FALSE, ...)
}
\arguments{
\item{x}{a class 'mable_reg' object return by functions such as \code{mable.ph} which contains 
\code{M}, \code{coefficients}, \code{p}, \code{m}, \code{x0}, \code{tau.n}, \code{tau} 
\code{lk}, \code{lr}.}

\item{y}{a new data.frame}

\item{newdata}{a new data.frame (ignored if \code{y} is included)}

\item{ntime}{number of evaluations of density, survival or cumulative distribution
curve to be plotted.}

\item{xlab}{x-axis label}

\item{which}{indicates which graphs to plot, options are 
"survival", "likelihood", "change-point", "density", or "all". If not "all", 
\code{which} can contain more than one options.}

\item{add}{logical add to an existing plot or not}

\item{...}{additional arguments to be passed to the base plot function}
}
\description{
Plot mathod for class 'mable_reg'
}
\author{
Zhong Guan <zguan@iusb.edu>
}
