% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vaal-Flow.r
\docType{data}
\name{Vaal.Flow}
\alias{Vaal.Flow}
\title{Vaal River Annual Flow Data}
\format{
The format is: int [1:65] 222 1094 452 1298 882 988 276 216 103 490 ...
}
\usage{
data(Vaal.Flow)
}
\description{
The annual flow data of Vaal River at Standerton as given by Table 1.1 of Linhart and Zucchini
(1986) give the flow in millions of cubic metres.
}
\examples{
data(Vaal.Flow)
}
\references{
Linhart, H., and Zucchini, W., \emph{Model Selection}, Wiley Series in Probability and Mathematical Statistics: Applied
Probability and Statistics, New York: John Wiley \& Sons Inc, 1986.
}
\keyword{datasets}
